requirements.txt
transformers
pandas
nltk
datasets
tokenizers
torch
nltk
spacy
pandas
numpy
rouge-score
tqdm
beautifulsoup4
python-dotenv
langchain
streamlit
tiktoken
unstructured
pdf2image
pdfminer
PyPDF2
flask
flask-cors
requests
nltk
python-dotenv
gunicorn
